Spreadsheet of topic w/ defining terms + scores
This is the spreadsheet we worked on w/ Prof. Larson. Note that, in the code, we are converting the scores to be 1-10 (weakest to strongest). Also note that all yellow rows are terms that weren't given a score by Larson. Those are supposed to be low priority, and I thought 0.5 might be an appropriate score.